+++ src/main/java/com/pepaproch/massmailmailer/db/entity/CampainAttachment.java
@@ -32,7 +32,7 @@ public class CampainAttachment implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "ID", nullable = false)
-    private BigDecimal id;
+    private Long id;
 
     @Basic(optional = false)
     @Column(name = "INDEX")
@@ -112,14 +112,14 @@ public class CampainAttachment implements Serializable {
     /**
      * @return the id
      */
-    public BigDecimal getId() {
+    public Long getId() {
         return id;
     }
 
     /**
      * @param id the id to set
      */
-    public void setId(BigDecimal id) {
+    public void setId(Long id) {
         this.id = id;
     }
